Update CraftTweaker, move classes, make command require OP
