+++ gradle.properties
@@ -22,7 +22,7 @@ mod_author=SkySom
 mod_id=packmode
 
 #CraftTweaker
-crafttweaker_version=9.1.192
+crafttweaker_version=9.1.197
 
 
 # Gradle
